Updated reducer structure
