+++ src/reducers/contextReducer.js
@@ -1,0 +1,15 @@
+const defaultState = 'PRE_GAME'
+const context = (state=defaultState, action) => {
+  switch (action.type) {
+    case "SWITCH_CONTEXT": {
+      state = action.payload
+      break;
+    }
+    default: {
+      break;
+    }
+  }
+  return state;
+}
+
+export default context;
