+++ src/reducers/index.js
@@ -1,17 +1,10 @@
-const defaultState = {
-  context: 'PRE_GAME'
-}
-const reducer = (state=defaultState, action) => {
-  switch (action.type) {
-    case "SWITCH_CONTEXT": {
-      state = {...state, context: action.payload}
-      break;
-    }
-    default: {
-      state = {...state}
-    }
-  }
-  return state;
-}
+import { combineReducers } from 'redux'
+import context from './contextReducer'
+import joinState from './joinReducer'
+
+const reducer = combineReducers({
+  context,
+  joinState
+})
 
 export default reducer;
